FIXED: ActLED didn't work with recent firmware
The ActLED didn't worked any more with the firmware from 2017-07-28 on.
This has been fixed.

CMemorySystem::GetCoherentPage() added

The coherent region was only one 64K page in size before.
Now it covers 1MB.

The A bit in SCTLR_EL1 will be explicitly cleared now to be sure.
Because this seems to be the default setting, this is not a change
in function.